shared/services/crm/requirements.txt
Django>=4.2.0
djangorestframework>=3.15.0
drf-nested-routers>=0.93.0
django-cors-headers>=4.3.0
psycopg2>=2.9.0
redis>=5.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-tenant-schemas>=1.9.0
django-filter>=23.0.0
Pillow>=10.0.0
python-decouple>=3.8.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0
pyjwt>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.5.0
django-health-check>=3.17.0
hvac==2.1.0  # HashiCorp Vault client for secrets management